+++ mirrord/layer-tests/src/intproxy.rs
@@ -65,6 +65,7 @@ impl TestIntProxy {
                 Default::default(),
                 Duration::from_secs(60),
                 &experimental_config,
+                mirrord_intproxy::session_monitor::MonitorTx::disabled(),
             );
             intproxy
                 .run(Duration::from_secs(15), Duration::from_secs(5))